pyproject.toml
[project]
name = "data-flux"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "asyncpg>=0.30.0",
    "openai-agents>=0.0.9",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.2",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
]

[dependency-groups]
dev = [
    "codespell>=2.4.1",
    "pre-commit>=4.2.0",
    "ruff>=0.11.4",
]
